requirements.txt
requests==2.18.3
sseclient-py==1.7
